requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
google-generativeai==0.8.3
aiosqlite==0.20.0
flask==3.1.0
flask-cors==5.0.0
requests==2.32.3
gunicorn==21.2.0
